Added support to build without docker image

+++ build.sh
@@ -4,17 +4,33 @@
 
 echo using GOPATH as $GOPATH
 
-docker run --rm -it -v "$GOPATH":/go -w /go/src/github.com/magneticio/vamp2cli dockercore/golang-cross:1.11.1 sh -c '
-for GOOS in darwin linux windows; do
-  for GOARCH in 386 amd64; do
-    echo "Building $GOOS-$GOARCH"
-    export GOOS=$GOOS
-    export GOARCH=$GOARCH
-    if [ "$GOOS" = "windows" ]; then
-      go get -u github.com/spf13/cobra
-    fi
-    go build -o bin/vamp2cli-$GOOS-$GOARCH
+if [ "$1" = "local" ]; then
+  for GOOS in darwin linux windows; do
+    for GOARCH in 386 amd64; do
+      echo "Building $GOOS-$GOARCH"
+      export GOOS=$GOOS
+      export GOARCH=$GOARCH
+      if [ "$GOOS" = "windows" ]; then
+        go get -u github.com/spf13/cobra
+      fi
+      go build -o bin/vamp2cli-$GOOS-$GOARCH
+    done
   done
-done
-'
+  unset GOOS
+  unset GOARCH
+else
+  docker run --rm -it -v "$GOPATH":/go -w /go/src/github.com/magneticio/vamp2cli dockercore/golang-cross:1.11.1 sh -c '
+  for GOOS in darwin linux windows; do
+    for GOARCH in 386 amd64; do
+      echo "Building $GOOS-$GOARCH"
+      export GOOS=$GOOS
+      export GOARCH=$GOARCH
+      if [ "$GOOS" = "windows" ]; then
+        go get -u github.com/spf13/cobra
+      fi
+      go build -o bin/vamp2cli-$GOOS-$GOARCH
+    done
+  done
+  '
+fi
 echo "Binaries can be found under bin directory"
